[FEATURE] allow more typolink.additionalParams for link generation
Currently the option "linkParams" is heavily bound to to the uid of the sqlMainTable
and hardcoded. It is not possible to add more additional params to the link.

This feature can be used to specify a special Controller and Action name instead
of the default one.

Example based on news extension is added to ext_typoscript_setup.txt